feat: health check 확인 후 트래픽 전환하도록 수정
- docker inspect로 새롭게 실행된 컨테이너의 health 상태 확인
- health check에 통과한 경우
  - nginx를 리로드하여 트래픽 전환
  - 구버전의 어플리케이션 종료
- health check에 실패한 경우
  - 신규버전에 문제가 있으므로 트래픽 전환하지 않음
  - 신규버전의 이미지를 삭제하여 디스크 공간 낭비 방지

+++ backend/deploy.sh
@@ -12,25 +12,47 @@ else
   WAS_STOP_PORT=3002
 fi
 
-DOCKER_COMPOSE_FILE="compose.$RUN_TARGET-deploy.yml"
+echo "The $STOP_TARGET version is currently running on the server. Starting the $RUN_TARGET version."
 
+DOCKER_COMPOSE_FILE="compose.$RUN_TARGET-deploy.yml"
 sudo docker-compose -f "$DOCKER_COMPOSE_FILE" pull
 sudo docker-compose -f "$DOCKER_COMPOSE_FILE" up -d
+sleep 20
 
-sleep 10
+echo "Starting health check for the new version of the application."
 
-NGINX_ID=$(sudo docker ps --filter "name=nginx" --quiet)
-NGINX_CONFIG="/etc/nginx/conf.d/default.conf"
+HEALTH_CHECK_PASSED=true
+RUN_CONTAINER_IDS=$(sudo docker ps --filter "name=$RUN_TARGET" --quiet)
 
-sudo docker exec $NGINX_ID /bin/bash -c "sed -i 's/was-$STOP_TARGET:$WAS_STOP_PORT/was-$RUN_TARGET:$WAS_RUN_PORT/' $NGINX_CONFIG"
-sudo docker exec $NGINX_ID /bin/bash -c "sed -i 's/signal-$STOP_TARGET:$((WAS_STOP_PORT + 1))/signal-$RUN_TARGET:$((WAS_RUN_PORT + 1))/' $NGINX_CONFIG"
-sudo docker exec $NGINX_ID /bin/bash -c "nginx -s reload"
+for CONTAINER_ID in $RUN_CONTAINER_IDS; do
+  HEALTH_STATUS=$(sudo docker inspect --format "{{.Health.Status}}" $CONTAINER_ID)
+  if [ "$HEALTH_STATUS" != "healthy" ]; then
+    HEALTH_CHECK_PASSED=false
+    break
+  fi
+done
 
-rm .env
+if [ "$HEALTH_CHECK_PASSED" = true ]; then
+  echo "Health check passed. Reloading nginx to transfer traffic from $STOP_TARGET to $RUN_TARGET."
+  
+  NGINX_ID=$(sudo docker ps --filter "name=nginx" --quiet)
+  NGINX_CONFIG="/etc/nginx/conf.d/default.conf"
 
-STOP_CONTAINER_ID=$(sudo docker ps --filter "name=$STOP_TARGET" --quiet)
-if [ -n "$STOP_CONTAINER_ID" ]; then
-  sudo docker stop $STOP_CONTAINER_ID
-  sudo docker rm $STOP_CONTAINER_ID
+  sudo docker exec $NGINX_ID /bin/bash -c "sed -i 's/was-$STOP_TARGET:$WAS_STOP_PORT/was-$RUN_TARGET:$WAS_RUN_PORT/' $NGINX_CONFIG"
+  sudo docker exec $NGINX_ID /bin/bash -c "sed -i 's/signal-$STOP_TARGET:$((WAS_STOP_PORT + 1))/signal-$RUN_TARGET:$((WAS_RUN_PORT + 1))/' $NGINX_CONFIG"
+  sudo docker exec $NGINX_ID /bin/bash -c "nginx -s reload"
+
+  echo "Terminating the $STOP_TARGET applications."
+
+  STOP_CONTAINER_ID=$(sudo docker ps --filter "name=$STOP_TARGET" --quiet)
+  if [ -n "$STOP_CONTAINER_ID" ]; then
+    sudo docker stop $STOP_CONTAINER_ID
+    sudo docker rm $STOP_CONTAINER_ID
+    sudo docker image prune -af
+  fi
+else
+  echo "Health check failed."
   sudo docker image prune -af
 fi
+
+rm .env
